Resolve regression test issues

+++ script-library/docker-debian.sh
@@ -52,7 +52,9 @@ if [ "${ENABLE_NONROOT_DOCKER}" = "true" ]; then
 
 set -e
 
-echo "(*) Ensuring ${NONROOT_USER} has access to ${SOURCE_SOCKET} via ${TARGET_SOCKET}"
+SOCAT_PATH_BASE=/tmp/vscr-dind-socat
+SOCAT_LOG=\${SOCAT_PATH_BASE}.log
+SOCAT_PID=\${SOCAT_PATH_BASE}.pid
 
 # Wrapper function to only use sudo if not already root
 sudoIf()
@@ -64,7 +66,14 @@ sudoIf()
     fi
 }
 
-SOCAT_LOG_PATH=/tmp/vscr-dind-socat.log
+# Log messages
+log()
+{
+    echo -e "[\$(date)] \$@" | sudoIf tee -a \${SOCAT_LOG} > /dev/null
+}
+
+echo -e "\n** \$(date) **" | sudoIf tee -a \${SOCAT_LOG} > /dev/null
+log "Ensuring ${NONROOT_USER} has access to ${SOURCE_SOCKET} via ${TARGET_SOCKET}"
 
 # If enabled, try to add a docker group with the right GID. If the group is root, 
 # fall back on using socat to forward the docker socket to another unix socket so 
@@ -72,7 +81,7 @@ SOCAT_LOG_PATH=/tmp/vscr-dind-socat.log
 if [ "${ENABLE_NONROOT_DOCKER}" = "true" ] && [ "${SOURCE_SOCKET}" != "${TARGET_SOCKET}" ] && [ "${NONROOT_USER}" != "root" ] && [ "${NONROOT_USER}" != "0" ]; then
     SOCKET_GID=\$(stat -c '%g' ${SOURCE_SOCKET})
     if [ "\${SOCKET_GID}" != "0" ]; then
-        echo "(*) Adding user to group with GID \${SOCKET_GID}."
+        log "Adding user to group with GID \${SOCKET_GID}."
         if [ "\$(cat /etc/group | grep :\${SOCKET_GID}:)" = "" ]; then
             sudoIf groupadd --gid \${SOCKET_GID} docker
         fi
@@ -81,13 +90,17 @@ if [ "${ENABLE_NONROOT_DOCKER}" = "true" ] && [ "${SOURCE_SOCKET}" != "${TARGET_
             sudoIf usermod -aG \${SOCKET_GID} ${NONROOT_USER}
         fi
     else
-        echo "(*) Enabling socket proxy."
-        date >> \${SOCAT_LOG_PATH}
-        echo "Proxying ${SOURCE_SOCKET} to ${TARGET_SOCKET} for ${NONROOT_USER}" >> \${SOCAT_LOG_PATH}
-        sudoIf rm -rf ${TARGET_SOCKET}
-        ((sudoIf socat UNIX-LISTEN:${TARGET_SOCKET},fork,mode=660,user=${NONROOT_USER} UNIX-CONNECT:${SOURCE_SOCKET}) 2>&1 >> \${SOCAT_LOG_PATH}) & > /dev/null
+        # Enable proxy if not already running
+        if [ ! -f "\${SOCAT_PID}" ] || ! ps -p \$(cat \${SOCAT_PID}) > /dev/null; then
+            log "Enabling socket proxy."
+            log "Proxying ${SOURCE_SOCKET} to ${TARGET_SOCKET} for vscode"
+            sudoIf rm -rf ${TARGET_SOCKET}
+            (sudoIf socat UNIX-LISTEN:${TARGET_SOCKET},fork,mode=660,user=${NONROOT_USER} UNIX-CONNECT:${SOURCE_SOCKET} 2>&1 | sudoIf tee -a \${SOCAT_LOG} > /dev/null & echo "\$!" | sudoIf tee \${SOCAT_PID} > /dev/null)
+        else
+            log "Socket proxy already running."
+        fi
     fi
-    echo "(*) Success"
+    log "Success"
 fi
 
 # Execute whatever commands were passed in (if any). This allows us
